+++ app/components/Hero2.tsx
@@ -1,61 +1,26 @@
 "use client";
 
-import React, { useEffect, useState, useRef } from "react";
+import React, { useRef } from "react";
 import {
   motion,
-  useAnimation,
-  AnimatePresence,
   useInView,
   easeOut,
   useScroll,
   useTransform,
 } from "framer-motion";
 import Hero3 from "./Hero3";
-
-interface ServiceCardProps {
-  title: string;
-  description: string;
-  icon: string;
-}
-
-const ServiceCard: React.FC<ServiceCardProps> = ({
-  title,
-  description,
-  icon,
-}) => {
-  return (
-    <motion.div
-      whileHover={{
-        y: -10,
-        boxShadow: "0 25px 50px rgba(0, 0, 0, 0.15)",
-        scale: 1.02,
-      }}
-      className="bg-white rounded-lg py-12 px-7 shadow-lg hover:shadow-xl transition-all duration-300 w-96 lg:w-[800px] lg:h-[400px] h-80 flex flex-col items-center justify-center"
-    >
-      <motion.div
-        className="w-20 h-20 bg-blue-100 rounded-full flex items-center justify-center mb-6"
-        animate={{
-          rotate: [0, 10, -10, 0],
-          scale: [1, 1.1, 1],
-        }}
-        transition={{
-          duration: 6,
-          repeat: Infinity,
-          ease: "easeInOut",
-        }}
-      >
-        <span className="text-3xl">{icon}</span>
-      </motion.div>
-      <h3 className="text-2xl font-semibold mb-3 text-center">{title}</h3>
-      <p className="text-gray-600 text-center text-base">{description}</p>
-    </motion.div>
-  );
-};
+import {
+  CheckCircle,
+  Video,
+  Headphones,
+  Mail,
+  User,
+  Building2,
+  Link as LinkIcon,
+  Send,
+} from "lucide-react";
 
 const Hero2: React.FC = () => {
-  const [count, setCount] = useState(0);
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const controls = useAnimation();
   const sectionRef = useRef(null);
   const isInView = useInView(sectionRef, { once: false, margin: "-100px" });
   const { scrollYProgress } = useScroll({
@@ -63,120 +28,90 @@ const Hero2: React.FC = () => {
     offset: ["start end", "end start"],
   });
 
-  const y = useTransform(scrollYProgress, [0, 1], [0, -30]);
+  const y1 = useTransform(scrollYProgress, [0, 1], [0, -50]);
+  const y2 = useTransform(scrollYProgress, [0, 1], [0, 50]);
   const scale = useTransform(scrollYProgress, [0, 0.5, 1], [1, 1.02, 1]);
 
-  const services = [
+  const features = [
     {
-      title: "Web Development",
-      description:
-        "Build responsive and scalable web applications with cutting-edge technologies.",
-      icon: "🌐",
+      label: "Automatic feature detection and highlighting",
+      icon: <CheckCircle size={18} className="text-purple-600" />,
     },
     {
-      title: "Mobile Development",
-      description:
-        "Create seamless mobile experiences for iOS and Android platforms.",
-      icon: "📱",
+      label: "Professional voiceover and background music",
+      icon: <Video size={18} className="text-purple-600" />,
     },
     {
-      title: "UI/UX Design",
-      description:
-        "Design intuitive and engaging user interfaces for better user experiences.",
-      icon: "🎨",
+      label: "Compelling call-to-action integration",
+      icon: <Headphones size={18} className="text-purple-600" />,
     },
   ];
 
-  useEffect(() => {
-    const animateCount = async () => {
-      await controls.start({
-        opacity: 1,
-        transition: { duration: 0.5 },
-      });
-
-      let start = 0;
-      const end = 1000;
-      const duration = 2000;
-      const increment = end / (duration / 16);
-
-      const timer = setInterval(() => {
-        start += increment;
-        if (start >= end) {
-          setCount(end);
-          clearInterval(timer);
-        } else {
-          setCount(Math.floor(start));
-        }
-      }, 16);
-
-      return () => clearInterval(timer);
-    };
-
-    animateCount();
-  }, [controls]);
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setCurrentIndex((prevIndex) => (prevIndex + 1) % services.length);
-    }, 5000);
-
-    return () => clearInterval(interval);
-  }, [services.length]);
-
-  const cardVariants = {
-    enter: { x: "100%", opacity: 0, scale: 0.8 },
-    center: { x: 0, opacity: 1, scale: 1 },
-    exit: { x: "-100%", opacity: 0, scale: 0.8 },
-  };
-
   return (
-    <section ref={sectionRef} className="relative overflow-hidden">
-      <motion.div
-        className="absolute bottom-1/4 right-10 w-16 h-16 bg-blue-200 rounded-full opacity-20"
-        animate={{
-          scale: [1, 1.3, 1],
-          x: [0, 30, 0],
-          rotate: [0, -180, -360],
-        }}
-        transition={{
-          duration: 10,
-          repeat: Infinity,
-          ease: "easeInOut",
-        }}
-      />
-
-      <div className="container mx-auto px-2 sm:px-4 relative z-10">
+    <>
+      <div
+        ref={sectionRef}
+        className="flex h-auto flex-col md:flex-row items-center justify-between p-6 md:p-12 bg-gradient-to-r from-white to-green-100 min-h-[400px] relative overflow-hidden"
+      >
         <motion.div
-          className="text-center mb-16"
-          initial={{ opacity: 0, y: 60, rotateX: -15 }}
+          className="absolute top-20 left-1/3 w-24 h-24  rounded-full opacity-20"
+          animate={{
+            scale: [1, 1.2, 1],
+            rotate: [0, 180, 360],
+            x: [0, 40, 0],
+          }}
+          transition={{
+            duration: 8,
+            repeat: Infinity,
+            ease: "linear",
+          }}
+        />
+        <motion.div
+          className="absolute bottom-20 right-1/4 w-20 h-20  rounded-full opacity-20"
+          animate={{
+            scale: [1, 1.3, 1],
+            y: [0, -30, 0],
+            rotate: [0, -180, -360],
+          }}
+          transition={{
+            duration: 10,
+            repeat: Infinity,
+            ease: "easeInOut",
+          }}
+        />
+
+        <motion.div
+          className="text-left max-w-lg ml-4 md:ml-16 relative z-10"
+          initial={{ opacity: 0, x: -60, rotateY: -15 }}
           animate={{
             opacity: isInView ? 1 : 0,
-            y: isInView ? 0 : 60,
-            rotateX: isInView ? 0 : -15,
+            x: isInView ? 0 : -60,
+            rotateY: isInView ? 0 : -15,
           }}
           transition={{
             duration: 0.8,
             ease: easeOut,
           }}
-          style={{ y }}
+          style={{ y: y1 }}
         >
-          <h2 className="text-3xl sm:text-5xl md:text-6xl mt-10 font-bold mb-4 text-gray-600 whitespace-nowrap">
-            Trusted by{" "}
-            <motion.span
-              className="text-[#6B46C1]"
-              initial={{ opacity: 0, scale: 0.8 }}
-              animate={controls}
-              whileHover={{
-                scale: 1.1,
-                textShadow: "0 0 20px rgba(107, 70, 193, 0.5)",
-              }}
-            >
-              {count}+
-            </motion.span>{" "}
-            Companies
-          </h2>
+          <motion.h1
+            className="text-4xl md:text-5xl font-extrabold text-gray-900 mb-6 leading-tight"
+            initial={{ opacity: 0, y: 60, rotateX: -15 }}
+            animate={{
+              opacity: isInView ? 1 : 0,
+              y: isInView ? 0 : 60,
+              rotateX: isInView ? 0 : -15,
+            }}
+            transition={{
+              duration: 0.8,
+              ease: easeOut,
+            }}
+          >
+            Ready to Transform Your{" "}
+            <span className="text-purple-600">Product Demos?</span>
+          </motion.h1>
           <motion.p
-            className="text-sm sm:text-xl md:text-2xl text-gray-700 max-w-full mx-auto whitespace-nowrap"
+            className="text-base md:text-lg text-gray-700 mb-8"
             initial={{ opacity: 0, y: 60 }}
             animate={{
               opacity: isInView ? 1 : 0,
@@ -188,46 +123,119 @@ const Hero2: React.FC = () => {
               delay: 0.2,
             }}
           >
-            See how businesses are using DemoAI to boost their conversations
+            Join thousands of companies using AI to create compelling demo
+            videos that convert. Start your free trial today.
           </motion.p>
+          <motion.ul
+            className="text-gray-700 space-y-6 text-base md:text-lg"
+            initial={{ opacity: 0 }}
+            animate={{ opacity: isInView ? 1 : 0 }}
+            transition={{
+              duration: 0.8,
+              ease: easeOut,
+              delay: 0.4,
+            }}
+          >
+            {features.map((item, index) => (
+              <motion.li
+                key={index}
+                className="flex items-start"
+                initial={{ opacity: 0, x: -20 }}
+                animate={{
+                  opacity: isInView ? 1 : 0,
+                  x: isInView ? 0 : -20,
+                }}
+                transition={{
+                  duration: 0.5,
+                  ease: easeOut,
+                  delay: 0.6 + index * 0.1,
+                }}
+                whileHover={{ x: 5, scale: 1.02 }}
+              >
+                <div className="w-8 h-8 rounded-md bg-purple-100 flex items-center justify-center">
+                  {item.icon}
+                </div>
+                <span>{item.label}</span>
+              </motion.li>
+            ))}
+          </motion.ul>
         </motion.div>
-
         <motion.div
-          className="relative w-full h-[400px] flex items-center justify-center overflow-hidden"
-          initial={{ opacity: 0, scale: 0.8, rotateZ: -10 }}
+          className="mt-12 md:mt-0 w-full max-w-md bg-[#C1A0FF] p-8 rounded-2xl shadow-xl relative z-10"
+          initial={{ opacity: 0, x: 60, rotateY: 15 }}
           animate={{
             opacity: isInView ? 1 : 0,
-            scale: isInView ? 1 : 0.8,
-            rotateZ: isInView ? 0 : -10,
+            x: isInView ? 0 : 60,
+            rotateY: isInView ? 0 : 15,
           }}
           transition={{
             duration: 0.8,
             ease: easeOut,
-            delay: 0.4,
+            delay: 0.3,
           }}
-          style={{ scale }}
+          style={{ y: y2, scale }}
+          whileHover={{
+            boxShadow: "0 25px 50px rgba(193, 160, 255, 0.3)",
+            y: -5,
+          }}
         >
-          <AnimatePresence initial={false}>
-            <motion.div
-              key={currentIndex}
-              variants={cardVariants}
-              initial="enter"
-              animate="center"
-              exit="exit"
-              transition={{ duration: 0.5, ease: "easeInOut" }}
-              className="absolute"
-            >
-              <ServiceCard
-                title={services[currentIndex].title}
-                description={services[currentIndex].description}
-                icon={services[currentIndex].icon}
+          <form className="space-y-6">
+            {[
+              { type: "email", placeholder: "Enter your Email", icon: <Mail size={18} /> },
+              { type: "text", placeholder: "Full Name", icon: <User size={18} /> },
+              { type: "text", placeholder: "Company", icon: <Building2 size={18} /> },
+              { type: "text", placeholder: "Product URL (Optional)",
+                icon: <LinkIcon size={18} /> },
+            ].map((input, index) => (
+              <motion.input
+                key={index}
+                type={input.type}
+                placeholder={input.placeholder}
+                className="w-full py-4 px-5 rounded-lg bg-purple-100 text-gray-700 text-base focus:outline-none focus:ring-2 focus:ring-purple-500"
+                whileFocus={{
+                  scale: 1.02,
+                  boxShadow: "0 0 20px rgba(147, 51, 234, 0.2)",
+                }}
+                initial={{ opacity: 0, y: 20 }}
+                animate={{
+                  opacity: isInView ? 1 : 0,
+                  y: isInView ? 0 : 20,
+                }}
+                transition={{
+                  duration: 0.5,
+                  ease: easeOut,
+                  delay: 0.5 + index * 0.1,
+                }}
               />
-            </motion.div>
-          </AnimatePresence>
+            ))}
+            <motion.button
+              type="submit"
+              className="w-full py-4 bg-white text-purple-600 font-semibold rounded-lg flex items-center justify-center gap-2 hover:bg-purple-100 transition"
+              whileHover={{
+                scale: 1.05,
+                boxShadow: "0 10px 25px rgba(0, 0, 0, 0.1)",
+                y: -2,
+              }}
+              whileTap={{ scale: 0.95 }}
+              initial={{ opacity: 0, y: 20 }}
+              animate={{
+                opacity: isInView ? 1 : 0,
+                y: isInView ? 0 : 20,
+              }}
+              transition={{
+                duration: 0.5,
+                ease: easeOut,
+                delay: 0.9,
+              }}
+            >
+              <Send size={18} />
+              Start Free Trial
+            </motion.button>
+          </form>
         </motion.div>
       </div>
       <Hero3 />
-    </section>
+    </>
   );
 };
